feat : 기존에 있었던 HomeDataRepository.kt, SanListRepository.kt, SanDetailRepository.kt 삭제하고 새로운 SanListRepository.kt를 data.repository 안에 생성함. 추가로 EventRepository.kt도 만들었다. 파이어베이스에서 같은 콜렉션에 있는 값을 한꺼번에 처리하면 좋을 거 같았는데 잘 실행되서 다행?
