+++ app/src/main/java/com/ing/offroader/data/model/userInfo/UserInfo.kt
@@ -1,0 +1,5 @@
+package com.ing.offroader.data.model.userInfo
+
+data class UserInfo(
+        var attendance : ArrayList<String>
+)
