+++ app/src/main/java/com/ing/offroader/data/repository/UserRepository.kt
@@ -1,0 +1,4 @@
+package com.ing.offroader.data.repository
+
+class UserRepository {
+}
