+++ app/src/main/java/com/ing/offroader/ui/fragment/home/HomeViewModel.kt
@@ -2,13 +2,21 @@ package com.ing.offroader.ui.fragment.home
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
+import com.ing.offroader.data.repository.EventRepository
+import com.ing.offroader.data.repository.SanListRepository
 
 
-class HomeViewModel(homeDataRepository: HomeDataRepository) : ViewModel() {
+class HomeViewModel(sanListRepository: SanListRepository, eventRepository: EventRepository) : ViewModel() {
 
-    private val repo : HomeDataRepository = homeDataRepository
+    private val sanListRepo : SanListRepository = sanListRepository
+    private val eventRepo : EventRepository = eventRepository
 
-    val recItems : LiveData<ArrayList<HomeUiState>> = repo.recItems
-    val eventItems : LiveData<ArrayList<HomeUiData.Fourth>> = repo.eventItems
+    val recItems : LiveData<ArrayList<HomeUiState>> = sanListRepo.recItems
+    val eventItems : LiveData<ArrayList<HomeUiData.Fourth>> = eventRepo.eventItems
+
+    fun takeAttendance() {
+
+    }
+
 
 }
