+++ app/src/main/res/layout/fragment_home.xml
@@ -50,6 +50,16 @@
         android:src="@drawable/apply_gradation_shadow_black"
         app:layout_constraintTop_toTopOf="@id/rv_home"/>
 
+    <Button
+        android:id="@+id/btn_attendance"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="20dp"
+        android:text="@string/home_attendance"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
+
     <!-- 커스텀 토스트?바 -->
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_customToast"
@@ -116,5 +126,7 @@
         app:layout_constraintBottom_toBottomOf="@id/tb_title"
         app:layout_constraintEnd_toEndOf="@id/tb_title"
         app:layout_constraintTop_toTopOf="@id/tb_title" />
+
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
