+++ app/src/main/res/values/strings.xml
@@ -79,5 +79,6 @@
     <string name="san_list_chip_alphabetical_order">가나다순</string>
     <string name="mountain_height_text">고도 :</string>
     <string name="signin_google">GOOGLE LOGIN</string>
+    <string name="home_attendance">출석체크</string>
 
 </resources>
